requirements.txt
python-dotenv
yandex-cloud-ml-sdk
requests
python-docx
Pillow
markdown
